requirements.txt
python-telegram-bot==21.5
pdfplumber==0.11.4
pytesseract==0.3.13
spacy==3.8.1
psycopg2==2.9.9
SQLAlchemy==2.0.34
filters==1.3.2
requests~=2.32.3
Flask~=3.0.3
torch~=2.4.1
python-dotenv~=1.0.1